pal-back-end/requirements.txt
aiofiles
fastapi
pydantic
python-dotenv    # for reading environment variables
uvicorn
loguru
redis
python-slugify
openai